feat: add devcontainer for the project (#143)
**Describe the pull request**
This pull request introduces a new feature: the addition of a
development container for the project. By providing a DevContainer,
we're creating a consistent development environment that is easy to set
up and use. This feature aims to streamline the onboarding process for
new developers and eliminate the "it works on my machine" problem by
ensuring everyone works in a standardized environment. By incorporating
this feature, we aim to improve the development process and ensure a
more reliable and uniform development experience.

**Checklist**
- [x] I have linked the relative issue to this pull request
- [x] I have made the modifications or added tests related to my PR
- [x] I have added/updated the documentation for my RP
- [x] I put my PR in Ready for Review only when all the checklist is checked