feat: formatting feature for postgresql query (#145)
**Relative Issues:** Resolve #106 

**Describe the pull request**
This pull request introduces a new feature: a formatting function for
PostgreSQL queries. This feature is designed to improve the readability
and maintainability of SQL queries by automatically applying consistent
formatting rules. The objective of this feature is to help developers
write clearer, more understandable SQL queries, making the code easier
to read, understand, and debug. By introducing this formatting feature,
we aim to enhance developer productivity and the overall quality of our
codebase.

**Checklist**

- [ ] I have linked the relative issue to this pull request
- [ ] I have made the modifications or added tests related to my PR
- [ ] I have added/updated the documentation for my RP
- [ ] I put my PR in Ready for Review only when all the checklist is
checked