chore: use koanf instead of viper to allow more possibility on configuration (#160)
**Describe the pull request**
This pull request involves switching from the Viper library to the Koanf
library in a Go application. The primary motivation behind this change
is to enhance the application's configuration possibilities. Viper,
although a comprehensive solution for configuration in Go applications,
has some limitations, particularly its case-insensitive behavior and a
large number of hardcoded dependencies, which result in a bigger binary
size compared to Koanf. For instance, a simple program using Viper is
significantly larger (about 313%) than its Koanf equivalent


**Checklist**

- [x] I have linked the relative issue to this pull request
- [x] I have made the modifications or added tests related to my PR
- [x] I have added/updated the documentation for my RP
- [x] I put my PR in Ready for Review only when all the checklist is
checked

**Additional context**
This switch to Koanf aims to reduce the application's binary size and
dependency footprint while maintaining or enhancing configuration
capabilities. The choice between Viper and Koanf largely depends on the
specific needs and constraints of the application in question.